requirements.txt
requests>=2.25.0
click>=8.0.0
rich>=10.0.0
tabulate>=0.8.9
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0